backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.1
fastapi==0.133.1
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
lap==0.5.13
MarkupSafe==3.0.2
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==12.0.0
polars==1.38.1
polars-runtime-32==1.38.1
psutil==7.2.2
pycocotools==2.0.11
pydantic==2.12.5
pydantic-core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
scipy==1.15.3
six==1.17.0
starlette==0.52.1
sympy==1.14.0

# PyTorch CPU versions compatible with Python 3.10
-f https://download.pytorch.org/whl/cpu/torch_stable.html
torch==2.0.1+cpu
torchvision==0.15.2+cpu
torchaudio==2.0.2+cpu

typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.18
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0